Stream-3/Full-Stack-Development/17.Create-A-Django-Based-Forum/7.Edit-And-Delete-A-Post/we_are_social/requirements.txt
arrow==0.10.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-debug-toolbar==1.8
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-js-asset==0.1.1
django-paypal==0.3.6
django-tinymce==2.6.0
idna==2.6
Pillow==3.1.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
stripe==1.62.1
urllib3==1.22
wrapt==1.10.11
reusable-blog-app==1.0.0
sqlparse==0.2.3
